Skip sorting if we don't want to sort.
The JS sort function @line 109 is not stable (order is not necessarily preserved for elements that compare 0) so when order == false and all elements compare 0, save the time and skip the sort call, so the original order is preserved.  

Also it seemed partly wrong for the compare function to return false instead of 0.  If the sort algorithm does false === 0 it may have issues.  This does seem unlikely though.